Updated print out.

+++ models/encoding/src/CharData.cpp
@@ -151,11 +151,11 @@ bool CharData::compare(
    }
    for ( int i = 0; i < min_ptr_char_size; ++i ) {
       if ( this->ptr_char[i] == data.ptr_char[i] ) {
-         msg << "this->ptr_char[" << i << "] (" << (int)this->ptr_char[i]
-             << ") == (" << (int)data.ptr_char[i] << ") data.ptr_char[" << i << "]\n";
+         msg << "this->ptr_char[" << i << "] (" << this->ptr_char[i]
+             << ") == (" << data.ptr_char[i] << ") data.ptr_char[" << i << "]\n";
       } else {
-         msg << "this->ptr_char[" << i << "] (" << (int)this->ptr_char[i]
-             << ") != (" << (int)data.ptr_char[i] << ") data.ptr_char[" << i << "]\n";
+         msg << "this->ptr_char[" << i << "] (" << this->ptr_char[i]
+             << ") != (" << data.ptr_char[i] << ") data.ptr_char[" << i << "]\n";
          equal_values = false;
       }
    }
